refactor(story response body data class): :recycle: replace data type of lat and lon
use data type `Double` instead of `Float` for easy to add google maps marker

+++ app/src/main/java/dev/achmadk/proasubmission1/models/StoryResponseBody.kt
@@ -12,8 +12,8 @@ data class StoryResponseBody (
     var description: String,
     var photoUrl: String,
     var createdAt: String,
-    var lat: Float? = null,
-    var lon: Float? = null,
+    var lat: Double? = null,
+    var lon: Double? = null,
     @PrimaryKey
     var id: String
 ): Parcelable
